+++ Assets/Tests/Devices/FiveFretGuitar/GuitarHeroGuitar/GuitarHeroGuitarTests.cs
@@ -7,10 +7,6 @@ namespace PlasticBand.Tests.Devices
     public sealed class GuitarHeroGuitarTests : PlasticBandTestFixture<GuitarHeroGuitar>
     {
         [Test]
-        public void TiltAndAccelerometerXAreEquivalent()
-            => CreateAndRun(_TiltAndAccelerometerXAreEquivalent);
-
-        [Test]
         public void GetTouchFretReturnsCorrectFrets()
             => CreateAndRun(_GetTouchFretReturnsCorrectFrets);
 
@@ -20,11 +16,6 @@ namespace PlasticBand.Tests.Devices
 
         // These must be named differently from the actual test methods, or else the input system test fixture
         // will fail to get the current method due to name ambiguity from reflection
-        public static void _TiltAndAccelerometerXAreEquivalent(GuitarHeroGuitar guitar)
-        {
-            Assert.That(guitar.accelX, Is.EqualTo(guitar.tilt));
-        }
-
         public static void _GetTouchFretReturnsCorrectFrets(GuitarHeroGuitar guitar)
         {
             FiveFretGuitarTests._GetFretReturnsCorrectFrets(guitar.GetTouchFret, guitar.GetTouchFret,
@@ -46,10 +37,6 @@ namespace PlasticBand.Tests.Devices
         protected abstract void SetSliderValue(ref TState state, byte value);
 
         [Test]
-        public void TiltAndAccelerometerXAreEquivalent()
-            => CreateAndRun(GuitarHeroGuitarTests._TiltAndAccelerometerXAreEquivalent);
-
-        [Test]
         public void GetTouchFretReturnsCorrectFrets()
             => CreateAndRun(GuitarHeroGuitarTests._GetTouchFretReturnsCorrectFrets);
 
@@ -68,22 +55,5 @@ namespace PlasticBand.Tests.Devices
             => CreateAndRun((guitar) =>
                 GuitarHeroSliderControlTests._HandlesGH5Slider(guitar, CreateState(), SetSliderValue,
                 guitar.touchGreen, guitar.touchRed, guitar.touchYellow, guitar.touchBlue, guitar.touchOrange));
-    }
-
-    public abstract class GuitarHeroGuitarTests_Accelerometer<TGuitar, TState> : GuitarHeroGuitarTests<TGuitar, TState>
-        where TGuitar : GuitarHeroGuitar
-        where TState : unmanaged, IInputStateTypeInfo
-    {
-        protected abstract void SetAccelerometerX(ref TState state, float value);
-        protected abstract void SetAccelerometerY(ref TState state, float value);
-        protected abstract void SetAccelerometerZ(ref TState state, float value);
-
-        [Test]
-        public void HandlesAccelerometers() => CreateAndRun((guitar) =>
-        {
-            RecognizesSignedAxis(guitar, CreateState(), guitar.accelX, SetAccelerometerX);
-            RecognizesSignedAxis(guitar, CreateState(), guitar.accelY, SetAccelerometerY);
-            RecognizesSignedAxis(guitar, CreateState(), guitar.accelZ, SetAccelerometerZ);
-        });
     }
 }